Merge Int to Short Adaptation to Main

+++ doku-solver/doku-solver/doku/Doku.cs
@@ -14,7 +14,7 @@ public class Doku{
         return nTab;
     }
 
-    protected bool IsSolved(int[,] tab){
+    protected bool IsSolved(short[,] tab){
         bool isSolved = true;
         for (int i = 0; i < tab.GetLength(0) && isSolved; i++){
             for (int j = 0; j < tab.GetLength(1) && isSolved; j++){
@@ -29,17 +29,17 @@ public class Doku{
         return IsSolved(grid.GetGrid());
     }
     
-    private readonly List<int> _rowPossibilities = new();
-    private readonly List<int> _columnPossibilities = new();
-    private readonly List<int> _sectionPossibilities = new();
-    private readonly List<int> _possibilities = new();
+    private readonly List<short> _rowPossibilities = new();
+    private readonly List<short> _columnPossibilities = new();
+    private readonly List<short> _sectionPossibilities = new();
+    private readonly List<short> _possibilities = new();
     
-    protected List<int> GetSlotPossibilities(int[,] tab, int row, int column){
+    protected List<short> GetSlotPossibilities(short[,] tab, int row, int column){
         _possibilities.Clear();
         GetRowPossibilities(tab, row);
         GetColumnPossibilities(tab, column);
         GetSectionPossibilities(tab, row, column);
-        foreach (int val in _rowPossibilities){
+        foreach (short val in _rowPossibilities){
             if(_columnPossibilities.Contains(val) && _sectionPossibilities.Contains(val)){
                 _possibilities.Add(val);
             }
@@ -47,12 +47,12 @@ public class Doku{
         return _possibilities;
     }
     
-    protected List<int> GetSlotPossibilities(Grid grid, Position position){
+    protected List<short> GetSlotPossibilities(Grid grid, Position position){
         return GetSlotPossibilities(grid.GetGrid(), position.Row, position.Column);
     }
-    private void GetColumnPossibilities(int[,] tab, int column){
+    private void GetColumnPossibilities(short[,] tab, int column){
         _columnPossibilities.Clear();
-        for(int i = 1; i <= tab.GetLength(0); i++){
+        for(short i = 1; i <= tab.GetLength(0); i++){
             _columnPossibilities.Add(i);
         }
         for (int i = 0; i < tab.GetLength(0); i++){
@@ -62,9 +62,9 @@ public class Doku{
         }
     }
     
-    private void GetRowPossibilities(int[,] tab, int row){
+    private void GetRowPossibilities(short[,] tab, int row){
         _rowPossibilities.Clear();
-        for(int i = 1; i <= tab.GetLength(1); i++){
+        for(short i = 1; i <= tab.GetLength(1); i++){
             _rowPossibilities.Add(i);
         }
         for (int i = 0; i < tab.GetLength(1); i++){
@@ -74,9 +74,9 @@ public class Doku{
         }
     }
     
-    private void GetSectionPossibilities(int[,] tab, int row, int column){
+    private void GetSectionPossibilities(short[,] tab, int row, int column){
         _sectionPossibilities.Clear();
-        for(int i = 1; i <= tab.GetLength(1); i++){
+        for(short i = 1; i <= tab.GetLength(1); i++){
             _sectionPossibilities.Add(i);
         }
         int squareSize = (int)Math.Sqrt(tab.GetLength(0));
